+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `/file_id` command to return photos file identifiers 
+
 ## [0.11.0] - 2023-03-21
 
 ### Changed